add ProductEventrepository
