+++ product-information/src/main/java/com/rewedigital/examples/msintegration/productinformation/product/ProductEventRepository.java
@@ -1,0 +1,7 @@
+package com.rewedigital.examples.msintegration.productinformation.product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductEventRepository extends JpaRepository<ProductEvent, String> {
+
+}
